Pull upnp devices out to separate class
This is a fairly large change, but it opens up the possibility of tapping in to
other upnp services that are on the tv. It uses closures to dynamically create
the upnp device methods from the scpd description.

I've tested as much as I can and all functionality seems to be retained.